+++ src/pages/Processes/Processes.jsx
@@ -14,9 +14,9 @@ import {
   Table,
   Content,
   ContentHeader,
-  Modal,
-  BackButton
+  Modal
 } from './styles';
+import BackButton from 'components/BackButton/BackButton';
 import api from 'services/api';
 import Button from 'components/Button/Button';
 import TextInput from 'components/TextInput/TextInput';
@@ -189,11 +189,7 @@ function Processes() {
   return (
     <Container>
       <div className="processes">
-        {flow && (
-          <BackButton onClick={() => navigate(-1)}>
-            <span>Voltar</span>
-          </BackButton>
-        )}
+        {flow && <BackButton />}
         <h1>Processos {flow ? '- ' + flow.name : ''}</h1>
         <div className="processSearch">
           <InputSearch
