+++ src/pages/Statistics/Statistics.jsx
@@ -2,14 +2,14 @@ import React, { useEffect, useState } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
 
 import api from 'services/api';
-import { Container, StagesArea, StageItem, BackButton } from './styles';
+import { Container, StagesArea, StageItem } from './styles';
+import BackButton from 'components/BackButton/BackButton';
 
 function Statistics() {
   const [stages, setStages] = useState([]);
   const [processes, setProcesses] = useState([]);
   const location = useLocation();
   const flow = location.state;
-  const navigate = useNavigate();
 
   function deleteStages(stagesResponse, stagesFlow, targetStages) {
     for (let stage of stagesResponse) {
@@ -58,9 +58,7 @@ function Statistics() {
   return (
     <>
       <Container>
-        <BackButton onClick={() => navigate(-1)}>
-          <span>Voltar</span>
-        </BackButton>
+        <BackButton />
         Estatística de fluxo
         <StagesArea>
           {stages.map((stage, index) => {
