+++ src/pages/StatisticsProcesses/Processes.jsx
@@ -12,7 +12,8 @@ import Button from 'components/Button/Button';
 import ModalBody from 'components/ModalBody/ModalBody';
 import TextInput from 'components/TextInput/TextInput';
 import ModalHeader from 'components/ModalHeader/ModalHeader';
-import { Container, InputSearch, AddProcess, BackButton } from './styles';
+import { Container, InputSearch, AddProcess } from './styles';
+import BackButton from 'components/BackButton/BackButton';
 
 function Processes() {
   const [searchTerm, setSearchTerm] = useState('');
@@ -30,7 +31,6 @@ function Processes() {
   const [stages, setStages] = useState([]);
   const [processes, setProcesses] = useState([]);
   const [currentStage, setCurrentStage] = useState('');
-  const navigate = useNavigate();
 
   const customStyles = {
     content: {
@@ -182,9 +182,7 @@ function Processes() {
   return (
     <Container>
       <div className="processes">
-        <BackButton onClick={() => navigate(-1)}>
-          <span>Voltar</span>
-        </BackButton>
+        <BackButton />
         <h1>Processos na etapa {currentStage ? '- ' + currentStage : ''}</h1>
         <div className="processSearch">
           <InputSearch
